+++ resources/views/classrooms/index.blade.php
@@ -19,16 +19,20 @@
     </div>
   </div>
 
-  @foreach ($user->classrooms->sortBy('state') as $classroom)
+  @foreach ($user->classrooms->sortByDesc('id')->sortBy('state') as $classroom)
 
   @if($classroom->state == 1)
-  @section('state')
-  <i class="far fa-hourglass-end"></i>
-  @endsection
+    @php
+      $icon = "<i class='far fa-hourglass-end' v-tippy :content='trans.get(\"classroom.finished\")'></i>";
+    @endphp
   @elseif($classroom->state == 2)
-  @section('state')
-  <i class="far fa-eye-slash"></i>
-  @endsection
+    @php
+      $icon= "<i class='far fa-eye-slash' v-tippy :content='trans.get(\"classroom.disabled\")'></i>";
+    @endphp
+  @else
+    @php
+      $icon = "";
+    @endphp
   @endif
   @php
   $theme = $classroom->theme;
@@ -37,8 +41,8 @@
     <div class="box w-100 h-100 d-flex is-relative rounded classroom-round text-center" @if($theme->type == 0) style="background-color: {{ $theme->color }}" @else style="background: url('/img/bg/thumb_{{ $theme->name }}');background-size: cover" @endif>
       <h3 class="is-size-4 has-text-light">
         @if($classroom->pivot->role == 2)
+        {!! $icon !!}
         <span class="admin">
-          @yield('state')
           <clone-classroom :classroom="{{ $classroom }}"></clone-classroom>
           <a href="/classroom/{{ $classroom->code }}/edit" class="has-text-white has-margin-x-2" {{ Popper::pop(__('menu.config')) }}><i class="far fa-cog"></i></a>
           <span {{ Popper::pop(__('classroom.admin')) }}>
@@ -49,16 +53,16 @@
           {{ $classroom->name }}
         </a>
         @elseif($classroom->pivot->role == 1)
+        {!! $icon !!}
         <span class="admin" {{ Popper::pop(__('classroom.teacher')) }}>
-          @yield('state')
           <i class="far fa-user-graduate"></i>
         </span>
         <a href="/classroom/{{ $classroom->code }}" class="has-text-light">
           {{ $classroom->name }}
         </a>
         @else
+        {!! $icon !!}
         <span class="admin" {{ Popper::pop(__('classroom.student')) }}>
-          @yield('state')
           <i class="far fa-user"></i>
         </span>
         <a href="/classroom/show/{{ $classroom->code }}" class="has-text-light">